+++ CHANGELOG.md
@@ -20,6 +20,20 @@ The format is based on **[Keep a Changelog](https://keepachangelog.com/en/1.1.0/
 - **Security**
   - (placeholder)
 
+## [1.1.0-beta.10] - 2025-10-30
+
+- **Added**
+  - (placeholder)
+
+- **Changed**
+  - (placeholder)
+
+- **Fixed**
+  - (placeholder)
+
+- **Security**
+  - (placeholder)
+
 ## [1.1.0-beta.9] - 2025-10-16
 
 - **Added**
@@ -327,7 +341,7 @@ The format is based on **[Keep a Changelog](https://keepachangelog.com/en/1.1.0/
 
 ---
 
-[Unreleased]: https://github.com/Plasius-LTD/voice/compare/v1.1.0-beta.9...HEAD
+[Unreleased]: https://github.com/Plasius-LTD/voice/compare/v1.1.0-beta.10...HEAD
 [1.0.0]: https://github.com/Plasius-LTD/voice/releases/tag/v1.0.0
 [1.0.4]: https://github.com/Plasius-LTD/voice/releases/tag/v1.0.4
 [1.0.5]: https://github.com/Plasius-LTD/voice/releases/tag/v1.0.5
@@ -345,3 +359,4 @@ The format is based on **[Keep a Changelog](https://keepachangelog.com/en/1.1.0/
 [1.1.0-beta.7]: https://github.com/Plasius-LTD/voice/releases/tag/v1.1.0-beta.7
 [1.1.0-beta.8]: https://github.com/Plasius-LTD/voice/releases/tag/v1.1.0-beta.8
 [1.1.0-beta.9]: https://github.com/Plasius-LTD/voice/releases/tag/v1.1.0-beta.9
+[1.1.0-beta.10]: https://github.com/Plasius-LTD/voice/releases/tag/v1.1.0-beta.10